Crate cpumask

CPU mask library Cpumasks provide a bitmap suitable for representing the set of CPUs in a system, one bit position per CPU number. In general, only nr_cpu_ids (<= NR_CPUS) bits are valid. Refering to cpumask_t in Linux. Reference: * https://man7.org/linux/man-pages/man2/sched_setaffinity.2.html * https://man7.org/linux/man-pages/man3/CPU_SET.3.html * https://elixir.bootlin.com/linux/v6.11/source/include/linux/cpumask_types.h

Structs

CpuMask

A compact array of bits which represents a set of physical CPUs, implemented based on bitmaps::Bitmap.

Iter

An iterator over the indices in a cpumask which are true.